Add PingPong Gameroom Content
